Update booking view for reviewing cancellations > #698
