--- client/src/components/Pages/Admin/Bookings/BookingView/index.js
@@ -1,1 +1,0 @@
-import React, { useState } from 'react';
